Supported mutating operators in nested expressions
